Toggler: Theme switcher fixed
The localStorage hook was removed due to the lack of usage inside the app. (Right now I am only using the localStorage API in the App component)

The whole idea of storing the toggler values was rewritten. (the values are stored in the custom hook in the app component and the setter function is sent down through the context)