Fix language flickering
Language flickering has been fixed.
The issue was simple. Because the "isEnglish" variable in the app component has three values (undefined, true, and false), it was possible that when it was undefined, it switched to the Hungarian language.
Now it shows "..." (temporary) for loading, and after it loads, it shows the CORRECT language.